+++ pl_bolts/models/self_supervised/__init__.py
@@ -24,6 +24,7 @@ from pl_bolts.models.self_supervised.cpc.cpc_module import CPCV2  # noqa: F401
 from pl_bolts.models.self_supervised.evaluator import SSLEvaluator  # noqa: F401
 from pl_bolts.models.self_supervised.moco.moco2_module import MocoV2  # noqa: F401
 from pl_bolts.models.self_supervised.simclr.simclr_module import SimCLR  # noqa: F401
+from pl_bolts.models.self_supervised.simsiam.simsiam_module import SimSiam  # noqa: F401
 from pl_bolts.models.self_supervised.ssl_finetuner import SSLFineTuner  # noqa: F401
 from pl_bolts.models.self_supervised.swav.swav_module import SwAV  # noqa: F401
 
@@ -34,6 +35,7 @@ __all__ = [
     "SSLEvaluator",
     "MocoV2",
     "SimCLR",
+    "SimSiam",
     "SSLFineTuner",
     "SwAV",
 ]